html/semantics/scripting-1/the-script-element/scriptexecute-events/029.html

<!DOCTYPE html>
<html>
  <title>afterscriptexecute for defer script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    events = [];
    var t = async_test();
    document.addEventListener("afterscriptexecute",
      function x(e) {
        events.push(e.target);
      }, false);
   addEventListener("DOMContentLoaded", function() {
    t.step(function(){assert_array_equals(events, [document.getElementsByTagName("script")[2], "next script ran", document.getElementsByTagName("script")[3]]);
           });
    t.done();
   });
 </script>
 <script defer src="include-4.js"></script>
